refactor(inbox): trim the change set
Ponytail pass: unexport a formatter with one caller in the same file,
keep list_inbox's user filter in one place instead of building the query
twice, hoist the SSE test wrapper that was pasted into two tests, and
share the adapter stub both inbox test files were declaring in full just
to satisfy the type.
